requirements.txt
bcrypt~=4.0.1
fastapi~=0.92.0
starlette~=0.25.0
pydantic~=1.10.5
pyyaml~=6.0
uvicorn~=0.20.0
passlib~=1.7.4
sqlalchemy~=2.0.5.post1
python-jose~=3.3.0
cryptography~=39.0.2
python-multipart~=0.0.6
onnx~=1.13.1
pytest~=7.2.2
httpx~=0.23.3